File: __canonicalize() fixes
If a directory doesn't exist, or if we can't read its files for some reason, catch the exception, and return the original name passed in. AIR seems to prefer keeping the original name passed in when something goes wrong.

If a file doesn't exist, don't return the all lower case version of the name. Keep the original casing of the passed in name.

Prefer an exact match on file name casing, and keep looking if a non-exact match with different casing is found. We should try to detect if the file system is case-sensitive or not in the future, though.