[Draft] LocalConnection: Fix messages larger than buffer size and improve message processing
Here we handle messages larger than the buffer. The approach on the native side  probably needs to be scrutinized. Windows handles message framing internally (optionally) and I took advantage of that here just to make things simpler, but because of that our read function return is kind of unorthodox. When introducing support for other platforms, we may want to handle message framing instead so as to unify the internal API across operating systems and make this easier to maintain. With that said, incremental improvements are fine.

I also made some more permanent changes to the way messages are serialized and parsed. Instead of using a fragile delimiter, now the implementation relies on pure binary which should make it a bit faster and improve stability.