TextEngine: fix missing default fonts on iOS and iPadOS
Arial is located in a different location on iPadOS than on iOS or macOS.

Added missing paths for Times New Roman and Courier New that don't contain spaces that affected both iOS and iPadOS.

Changed order of Helvetica and Times fallbacks for Arial and Times New Roman. It alternated previously, which made it more likely to have a mix of Helvetica and Arial, for example. Now, it exhaustively searches for one before checking the fallback, so that it's more likely that there won't be a mix. In other words, check for Arial in all possible locations before trying Helvetica as a fallback.