Graphics: fix crash when drawing to BitmapData after clear() and lineStyle() only (closes #2777)
After clear(), graphics.__width and graphics.__height were still > 0. This is because Graphics.__update() was not updating these values when returning early because bounds.width or bounds.height <= 0. Later in __update() method, when scaleX/scaleY is applied, it resets __width and __height to 0 and returns, so I copied that to the first early return to match behavior.