TExtField: fix setTextInputRect being ignored on macOS (maybe elsewhere too) because it is called before setting window.textInputEnabled to true
When SDL ignores the text input rect, it uses 0,0 instead. On macOS, this results in any overlays (such as the caps lock indicator) appearing in the top left corner, instead of relative to the rectangle. Typing would fix the position temporarily while the text field still had focus, but changing focus to another text field, or to something else and back to the original text field would revert back at 0,0.